Add 'are you sure' prompt to delete step in frontend
