+++ client/src/styles/modals.js
@@ -1,0 +1,13 @@
+const style = {
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    transform: 'translate(-50%, -50%)',
+    width: 400,
+    bgcolor: '#eee',
+    borderRadius: '10px',
+    boxShadow: 24,
+    p: 4,
+};
+
+export default style;
